Finished 3-2
Experimented with mixing the different options for defining types. I think I'm a bit more of a fan of a verbose style that declares the full signature in a classical functional style using implication (->) and then defining the variables via lambdas and not assumes. (Wish you could use implication and name the propositions altogether in the type signature like in Idris)